requirements.txt
fastapi==0.77.1
faiss-gpu==1.7.2
numpy==1.21.6
pandas==1.3.5
pathlib==1.0.1
pip-tools==6.2.0
sentence-transformers==2.2.0
datasets==2.2.1
torch @ https://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp37-cp37m-linux_x86_64.whl
uvicorn==0.13.1
gunicorn==20.1.0
transformers==4.19.2
sentencepiece==0.1.96
tqdm==4.64.0